cleaned up the controllers, can get all shots now

+++ server/controller/animals.go
@@ -13,44 +13,18 @@ import (
 )
 
 func AnimalController(w http.ResponseWriter, r *http.Request) {
-	w.Header().Set("Access-Control-Allow-Origin", "*")
-	w.Header().Set("Access-Control-Allow-Methods", "GET, POST, PUT, DELETE")
-	w.Header().Set("Access-Control-Allow-Headers", "Content-Type, Authorization")
-
-	if r.URL.Path == "/organizations/animals" {
-		switch r.Method {
-		case http.MethodPost:
-			w.Header().Set("Content-Type", "application/json")
-
-			hasId, id := checkForBodyItem("id", w, r)
-			if hasId {
-				if err := json.NewEncoder(w).Encode(database.GetAnimal(id)); err != nil {
-					http.Error(w, "unable to encode response", http.StatusInternalServerError)
-				}
-				return
-			}
-			fmt.Fprintf(w, "Body does not have an id!")
-
-		case http.MethodGet:
-			w.Header().Set("Content-Type", "application/json")
-			if err := json.NewEncoder(w).Encode(database.GetAllAnimals()); err != nil {
-				http.Error(w, "unable to encode response", http.StatusInternalServerError)
-			}
-			return
-		}
-	}
+	SetHeader(w)
 
 	if r.URL.Path == "/animals" {
 
 		switch r.Method {
 		case http.MethodPost:
-
 			w.Header().Set("Content-Type", "application/json")
 			fmt.Println("The else in method post was called")
 
 			animal := GetAnimalFromBody(w, r)
 
-			// database.InsertAnimal(animal)
+			database.InsertAnimal(animal)
 			fmt.Fprintf(w, "Animal created Successfully!: %v\n", animal)
 		case http.MethodPut:
 			w.Header().Set("Content-Type", "application/json")
@@ -62,10 +36,14 @@ func AnimalController(w http.ResponseWriter, r *http.Request) {
 		case http.MethodDelete:
 			w.Header().Set("Content-Type", "application/json")
 
-			_, id := checkForBodyItem("id", w, r)
+			hasId, id := checkForBodyItem("id", w, r)
 
-			database.DeleteAnimal(id)
-			fmt.Fprintf(w, "Animal was removed successfully")
+			if hasId {
+				database.DeleteAnimal(id)
+				fmt.Fprintf(w, "Animal was removed successfully")
+				return
+			}
+			fmt.Fprintf(w, "Id was not found in body!")
 		}
 	}
 }
@@ -102,7 +80,6 @@ func GetAnimalFromBody(w http.ResponseWriter, r *http.Request) database.NewAnima
 		return database.NewAnimal{}
 	}
 
-	fmt.Printf("Animal inside GAFB is %v\n", animal)
 	return animal
 }
